Further specified child experiments.
Modified columns in experiments so most allow null, and none have defaults. Made that migration irreversible for no particular reason, except that it's a tricky down-migration to write.

Experiments created on matrices with grandchildren will now automatically create sub-experiments without duplicating the parent experiments. The only real purpose of this is to save some DB space when we start accumulating lots and lots of experiments.