Sub-classed Experiment with Integrator, created Integrand model.
Added basic views, have not checked to see if they work. Migration for integrand has not been run, and may need some more columns.

Also added to_r method for String, for converting to rational numbers. This is used by Integrand, which has weightn and weightd, which combine to create a weight rational. This ensures that all Integrands add up to exactly 1.