Added a progress bar and children multiplot for two-stage cross-validations.
Cleaned up a lot of the experiment listings. Converted some timestamps to distance in words.

Added progress bars that show how many experiments have been run on a matrix or an experiment with children.

Added a function to experiment, aucs_by_column_with_children, which allows children to be plotted along with the properly-sorted parent. This shows correlations for when the exact same analysis performed is performed on parents and children.

Also managed to make it so the points are very small.

I may have also fixed a few bugs involving errors when displaying plots. Js files were missing.